-always finding flag :) -bug fixes

+++ src/main/java/com/bartosz/lewandowski/ai/astar/models/Edge.java
@@ -1,10 +1,10 @@
 package com.bartosz.lewandowski.ai.astar.models;
 
 public class Edge {
-    public final double cost;
+    public final int cost;
     public final Node target;
 
-    public Edge(Node targetNode, double costVal) {
+    public Edge(Node targetNode, int costVal) {
         target = targetNode;
         cost = costVal;
     }
